Fix nfl().
The function wasn't working. It was treating $xml as a string containing
XML instead of a filename when checking to see if a team had any games
available to display.

Also, changed it so that now you have to specify the teams when running
`nfl` rather than hard coding them into .bashrc.